FIXED  : Remember position not working with some DVDs ADDED  : Display stat recognized Intel ClearVideo DXVA modes CHANGED: MpcVideoDec standalone filter can connect to MediaPortal
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@806 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/subpic/DX7SubPic.cpp
@@ -270,8 +270,8 @@ bool CDX7SubPicAllocator::Alloc(bool fStatic, ISubPic** ppSubPic)
 	if(m_fPow2Textures)
 	{
 		ddsd.dwWidth = ddsd.dwHeight = 1;
-		while(ddsd.dwWidth < m_maxsize.cx) ddsd.dwWidth <<= 1;
-		while(ddsd.dwHeight < m_maxsize.cy) ddsd.dwHeight <<= 1;
+		while(ddsd.dwWidth < (DWORD)m_maxsize.cx) ddsd.dwWidth <<= 1;
+		while(ddsd.dwHeight < (DWORD)m_maxsize.cy) ddsd.dwHeight <<= 1;
 	}
